REFACTOR :: code format 정리
- 뒤로가기 추가
- post componenet 누르면 들어가도록
