+++ src/components/Post/detail/PostDetailComment.js
@@ -1,5 +1,5 @@
-import { useState } from "react";
-import { StyleSheet, Text, View } from "react-native";
+import {useState} from "react";
+import {StyleSheet, View} from "react-native";
 import PostDetailCommentItem from "./PostDetailCommentItem";
 
 const PostDetailComment = () => {
