+++ src/routes/MainNavigation.js
@@ -1,6 +1,7 @@
 import ChatLiveScreen from "../screens/ChatLiveScreen";
 import ChatListScreen from "../screens/ChatListScreen";
 import TabNavigation from "./TabNavigation";
+import PostDetailScreen from "../screens/PostDetailScreen";
 
 
 export default function MainNavigation({Stack, socket}) {
@@ -17,6 +18,7 @@ export default function MainNavigation({Stack, socket}) {
             <Stack.Screen name="ChatLiveScreen">
                 {props => <ChatLiveScreen {...props} socket={socket}/>}
             </Stack.Screen>
+            <Stack.Screen name="PostDetailScreen" component={PostDetailScreen}/>
         </Stack.Navigator>
     )
 };
